[kafkazk] break out zookeeper errors
